# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge RAG API
  version: 1.2.0
  description: |
    基于当前代码实现整理的接口规范，覆盖登录鉴权、注册码管理、文档上传管理、公开文档浏览与 RAG 问答接口。

    说明：
    - 文档管理接口使用 Session Cookie 鉴权，登录成功后服务端会写入 `JSESSIONID`。
    - 公开文档接口（`/documents/public/**`）无需鉴权，所有限流保护。
    - RAG 问答接口保持匿名可访问。
    - `/auth/register` 只能使用管理员创建的一次性注册码注册。
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /documents/upload:
    post:
      tags:
        - Documents
      summary: 上传并处理文档
      operationId: uploadDocument
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 支持 `.pdf`、`.txt`、`.doc`、`.docx`
      responses:
        '200':
          description: 文档处理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: 文件为空、文件名缺失或文件格式不支持
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: 文档解析、切块或向量入库失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents:
    get:
      tags:
        - Documents
      summary: 获取已上传文档列表
      operationId: listDocuments
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 文档列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: 查询失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{documentId}:
    delete:
      tags:
        - Documents
      summary: 删除指定文档
      operationId: deleteDocument
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDeleteResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: 未找到对应文档
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 删除失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/public:
    get:
      tags:
        - Public Documents
      summary: 获取公开文档列表
      operationId: listPublicDocuments
      description: 返回所有文档的元数据列表，无需鉴权。受 IP 限流保护（30 次/分钟）。
      responses:
        '200':
          description: 文档列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocumentListResponse'
        '429':
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 查询失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/public/{documentId}:
    get:
      tags:
        - Public Documents
      summary: 获取公开文档详情
      operationId: getPublicDocumentDetail
      description: 返回文档元数据及全部文本段落，无需鉴权。受 IP 限流保护（30 次/分钟）。
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 文档详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocumentDetailResponse'
        '404':
          description: 文档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 查询失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/public/{documentId}/download:
    get:
      tags:
        - Public Documents
      summary: 下载文档原文件
      operationId: downloadPublicDocument
      description: 下载文档原始文件（如 PDF、TXT、DOC、DOCX），无需鉴权。受 IP 限流保护（10 次/分钟）。
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 文件二进制流
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
                format: binary
        '404':
          description: 文档或文件不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 下载失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    UnauthorizedError:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 未登录
            message: 请先登录后再访问该接口
            timestamp: '2026-03-29T18:00:00'
...
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      description: 文档唯一标识
      schema:
        type: string
...
  schemas:
...
    DocumentResponse:
      type: object
      required:
        - message
        - segmentCount
      properties:
        documentId:
          type: string
          nullable: true
        filename:
          type: string
          nullable: true
        message:
          type: string
        segmentCount:
          type: integer
          format: int32
          minimum: 0
    DocumentListResponse:
      type: object
      required:
        - documents
        - total
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentListItemResponse'
        total:
          type: integer
          format: int32
          minimum: 0
    DocumentListItemResponse:
      type: object
      required:
        - documentId
        - filename
        - segmentCount
      properties:
        documentId:
          type: string
        filename:
          type: string
        segmentCount:
          type: integer
          format: int32
          minimum: 0
    DocumentDeleteResponse:
      type: object
      required:
        - documentId
        - deletedSegments
        - message
      properties:
        documentId:
          type: string
        filename:
          type: string
          nullable: true
        deletedSegments:
          type: integer
          format: int32
          minimum: 0
        message:
          type: string
    PublicDocumentSegment:
      type: object
      required:
        - chunkIndex
        - text
      properties:
        chunkIndex:
          type: integer
          format: int32
        text:
          type: string
    PublicDocumentListItem:
      type: object
      required:
        - documentId
        - filename
        - title
        - category
        - documentTime
        - keywords
        - segmentCount
      properties:
        documentId:
          type: string
        filename:
          type: string
        title:
          type: string
        category:
          type: string
        documentTime:
          type: string
        keywords:
          type: string
        segmentCount:
          type: integer
          format: int32
          minimum: 0
    PublicDocumentListResponse:
      type: object
      required:
        - documents
        - total
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/PublicDocumentListItem'
        total:
          type: integer
          format: int32
          minimum: 0
    PublicDocumentDetailResponse:
      type: object
      required:
        - documentId
        - filename
        - title
        - category
        - documentTime
        - keywords
        - segmentCount
        - segments
      properties:
        documentId:
          type: string
        filename:
          type: string
        title:
          type: string
        category:
          type: string
        documentTime:
          type: string
        keywords:
          type: string
        segmentCount:
          type: integer
          format: int32
          minimum: 0
        segments:
          type: array
          items:
            $ref: '#/components/schemas/PublicDocumentSegment'
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...